refactor(search-view): extract useProfileScope hook
